requirements.txt
antlr4-python3-runtime==4.7.2
colour==0.1.5
playsound==1.2.2
pysynth==2.3
typing==3.6.6
PySynth==2.3
pydub==0.23.1
